rename package detectionStrategie to HandDetectors
